update in types.d.ts
