+++ assets/src/modules/gallery/gallery-module.ts
@@ -1,25 +1,7 @@
-import { AuthConfig } from './../../upload-utils.ts';
+import type { MediaGalleryOptions } from './../../types.d.ts';
 import GalleryModal from './gallery-modal.ts';
 import Quill from 'quill';
 import 'styles/gallery/gallery.css';
-
-type MediaGalleryOptions = {
-    uploadEndpoint: string;
-    listEndpoint: string;
-    searchEndpoint: string;
-    icon: string;
-    buttonTitle: string;
-    uploadTitle: string;
-    messageLoadingOption: string;
-    messageNextPageOption: string;
-    messagePrevPageOption: string;
-    messageErrorOption: string;
-    messageNoImageOption: string;
-    messageSearchPlaceholderOption: string;
-    authConfig?: AuthConfig;
-    jsonResponseFilePath?: string;
-    uploadStrategy?: 'form' | 'json';
-}
 
 export default class GalleryModule {
     private quill: Quill;
@@ -32,7 +14,7 @@ export default class GalleryModule {
             uploadEndpoint: options.uploadEndpoint || '',
             listEndpoint: options.listEndpoint || '',
             searchEndpoint: options.searchEndpoint || '',
-            icon: options.icon,
+            icon: options.icon || '',
             buttonTitle: options.buttonTitle || '',
             uploadTitle: options.uploadTitle || '',
             messageLoadingOption: options.messageLoadingOption || '',
@@ -63,8 +45,12 @@ export default class GalleryModule {
         const button = document.createElement('button');
         button.type = 'button';
         button.classList.add('ql-gallery');
-        button.innerHTML = this.options.icon;
-        button.title = this.options.buttonTitle;
+        if (this.options.icon != null) {
+            button.innerHTML = this.options.icon;
+        }
+        if (this.options.buttonTitle != null) {
+            button.title = this.options.buttonTitle;
+        }
         button.addEventListener('click', () => this.open());
 
         const group = document.createElement('span');
